requirements.txt
PySide6
PyYAML
configparser